+++ tests/test_visitors/test_ast/test_complexity/test_counts/test_condition_counts.py
@@ -8,6 +8,7 @@ from wemake_python_styleguide.visitors.ast.complexity.counts import (
 )
 
 empty_module = ''
+assignment = 'some = x > y or None'
 
 condition_with_single_if = """
 if 4 > 2 and 3 / 2 == 1.5:
@@ -44,9 +45,26 @@ while True and 1 == 1:
     print(1)
 """
 
+# Real examples:
+
+complex_assignment = """
+some = zero and first or (second and last) or default()
+"""
+
+complex_condition = """
+if x == x1 and y == y1 and z == z1 or v == v1 or last():
+    ...
+"""
+
+complex_while = """
+while (x > x1 or y < y1) or (small(z) and v) or last():
+    ...
+"""
+
 
 @pytest.mark.parametrize('code', [
     empty_module,
+    assignment,
     condition_with_single_if,
     condition_with_single_if_multiline,
     condition_with_several_ifs,
@@ -57,7 +75,10 @@ while True and 1 == 1:
     while_with_condition,
 ])
 def test_module_condition_counts_normal(
-    assert_errors, parse_ast_tree, code, default_options,
+    assert_errors,
+    parse_ast_tree,
+    code,
+    default_options,
 ):
     """Testing that conditions in a module work well."""
     tree = parse_ast_tree(code)
@@ -69,6 +90,7 @@ def test_module_condition_counts_normal(
 
 
 @pytest.mark.parametrize('code', [
+    assignment,
     condition_with_single_if,
     condition_with_single_if_multiline,
     condition_with_several_ifs,
@@ -78,13 +100,41 @@ def test_module_condition_counts_normal(
     while_with_condition,
 ])
 def test_module_condition_counts_violation(
-    assert_errors, parse_ast_tree, code, options,
+    monkeypatch,
+    assert_errors,
+    assert_error_text,
+    parse_ast_tree,
+    code,
+    default_options,
 ):
     """Testing that violations are raised when reaching max value."""
     tree = parse_ast_tree(code)
 
-    option_values = options(max_conditions=1)
-    visitor = ConditionsVisitor(option_values, tree=tree)
+    monkeypatch.setattr(ConditionsVisitor, '_max_conditions', 1)
+    visitor = ConditionsVisitor(default_options, tree=tree)
     visitor.run()
 
     assert_errors(visitor, [TooManyConditionsViolation])
+    assert_error_text(visitor, '2')
+
+
+@pytest.mark.parametrize('code', [
+    complex_assignment,
+    complex_condition,
+    complex_while,
+])
+def test_module_condition_real_config(
+    assert_errors,
+    assert_error_text,
+    parse_ast_tree,
+    default_options,
+    code,
+):
+    """Testing that violations are raised when reaching max value."""
+    tree = parse_ast_tree(code)
+
+    visitor = ConditionsVisitor(default_options, tree=tree)
+    visitor.run()
+
+    assert_errors(visitor, [TooManyConditionsViolation])
+    assert_error_text(visitor, '5')