+++ tests/test_visitors/test_ast/test_complexity/test_function/test_elifs.py
@@ -7,40 +7,88 @@ from wemake_python_styleguide.visitors.ast.complexity.function import (
     TooManyElifsViolation,
 )
 
+function_with_one_elif = """
+def test_module():
+    if 1 > 2:
+        ...
+    elif 2 > 3:
+        ...
+    else:
+        ...
+"""
+
+function_with_two_elifs = """
+def test_module():
+    if 1 > 2:
+        ...
+    elif 2 > 3:
+        ...
+    else:
+        ...
+"""
+
+function_with_three_elifs = """
+def test_module():
+    if 1 > 2:
+        ...
+    elif 2 > 3:
+        ...
+    else:
+        ...
+"""
+
 function_with_elifs = """
 def test_module():
     if 1 > 2:
-        print(1)
+        ...
     elif 2 > 3:
-        print(2)
+        ...
     elif 3 > 4:
-        print(3)
+        ...
+    elif 4 > 5:
+        ...
+    elif 5 > 6:
+        ...
     else:
-        print(4)
+        ...
+"""
+
+function_with_elifs_without_else = """
+def test_module():
+    if 1 > 2:
+        ...
+    elif 2 > 3:
+        ...
+    elif 3 > 4:
+        ...
+    elif 4 > 5:
+        ...
+    elif 5 > 6:
+        ...
 """
 
 function_with_ifs = """
 def test_module():
     if True:
-        print(1)
+        ...
     if 2 > 3:
-        print(2)
+        ...
     if 3 > 4:
-        print(3)
+        ...
 """
 
 function_with_raw_if = """
 def function():
     if 1 == 2:
-        print(1)
+        ...
 """
 
 function_with_if_else = """
 def function(param):
     if param == 2:
-        print(1)
+        ...
     else:
-        print(2)
+        ...
 """
 
 function_with_ternary = """
@@ -50,7 +98,9 @@ def with_ternary(some_value):
 
 
 @pytest.mark.parametrize('code', [
-    function_with_elifs,
+    function_with_one_elif,
+    function_with_two_elifs,
+    function_with_three_elifs,
     function_with_ifs,
     function_with_raw_if,
     function_with_if_else,
@@ -63,7 +113,7 @@ def test_elif_correct_count(
     default_options,
     mode,
 ):
-    """Testing that all `if`/`elif`/`else` stuff is allowed."""
+    """Testing that all `if`/`elif`/`else` is allowed."""
     tree = parse_ast_tree(mode(code))
 
     visitor = FunctionComplexityVisitor(default_options, tree=tree)
@@ -74,21 +124,20 @@ def test_elif_correct_count(
 
 @pytest.mark.parametrize('code', [
     function_with_elifs,
+    function_with_elifs_without_else,
 ])
 def test_elif_incorrect_count(
     assert_errors,
     assert_error_text,
     parse_ast_tree,
     code,
-    options,
-    mode,
+    default_options,
 ):
-    """Testing that incorrect number of `elif` stuff is restricted."""
-    tree = parse_ast_tree(mode(code))
+    """Testing that incorrect number of `elif` is restricted."""
+    tree = parse_ast_tree(code)
 
-    option_values = options(max_elifs=1)
-    visitor = FunctionComplexityVisitor(option_values, tree=tree)
+    visitor = FunctionComplexityVisitor(default_options, tree=tree)
     visitor.run()
 
     assert_errors(visitor, [TooManyElifsViolation])
-    assert_error_text(visitor, '2')
+    assert_error_text(visitor, '4')