+++ wemake_python_styleguide/visitors/ast/complexity/__init__.py
@@ -1,3 +1,1 @@
 # -*- coding: utf-8 -*-
-
-# TODO: refactor "complexity" visitors to be regular visitors