+++ wemake_python_styleguide/visitors/presets/complexity.py
@@ -9,7 +9,7 @@ from wemake_python_styleguide.visitors.ast.complexity import (
 )
 
 #: Used to store all complexity related visitors to be later passed to checker:
-COMPLEXITY_PRESET = (
+COMPLEXITY_PRESET = (  # TODO: remove this preset
     function.FunctionComplexityVisitor,
     jones.JonesComplexityVisitor,
     nested.NestedComplexityVisitor,